@@ -4,13 +4,16 @@
 	"description": "Check if a Buffer/Uint8Array is a GIF image",
 	"license": "MIT",
 	"repository": "sindresorhus/is-gif",
+	"funding": "https://github.com/sponsors/sindresorhus",
 	"author": {
 		"name": "Sindre Sorhus",
 		"email": "sindresorhus@gmail.com",
-		"url": "sindresorhus.com"
+		"url": "https://sindresorhus.com"
 	},
+	"type": "module",
+	"exports": "./index.js",
 	"engines": {
-		"node": ">=6"
+		"node": "^12.20.0 || ^14.13.1 || >=16.0.0"
 	},
 	"scripts": {
 		"test": "xo && ava"
@@ -22,8 +25,6 @@
 		"gif",
 		"graphics",
 		"image",
-		"img",
-		"pic",
 		"picture",
 		"photo",
 		"type",
@@ -35,11 +36,8 @@
 		"buffer",
 		"uint8array"
 	],
-	"dependencies": {
-		"file-type": "^10.4.0"
-	},
 	"devDependencies": {
-		"ava": "^0.25.0",
-		"xo": "^0.23.0"
+		"ava": "^3.15.0",
+		"xo": "^0.44.0"
 	}
 }
