@@ -1,12 +1,9 @@
-'use strict';
-const fileType = require('file-type');
-
-module.exports = input => {
-	const match = fileType(input);
-
-	if (!match) {
+export default function isGif(buffer) {
+	if (!buffer || buffer.length < 3) {
 		return false;
 	}
 
-	return match.ext === 'gif';
-};
+	return buffer[0] === 0x47
+		&& buffer[1] === 0x49
+		&& buffer[2] === 0x46;
+}
