@@ -1,6 +1,6 @@
-import fs from 'fs';
+import fs from 'node:fs';
 import test from 'ava';
-import isGif from '.';
+import isGif from './index.js';
 
 const check = filename => isGif(fs.readFileSync(filename));
 
@@ -8,18 +8,3 @@ test('detect GIF from Buffer', t => {
 	t.true(check('fixture.gif'));
 	t.false(check('fixture.png'));
 });
-
-test('validate the input argument type', t => {
-	t.throws(() => {
-		isGif('x');
-	}, /Expected the `input` argument to be of type `Uint8Array`/);
-
-	t.notThrows(() => {
-		isGif(Buffer.from('x'));
-	});
-
-	t.notThrows(() => {
-		isGif(new Uint8Array());
-	});
-});
-
